+++ libft/srcs/ft_strsplit.c
@@ -53,7 +53,7 @@ char				**ft_strsplit(const char *s, char c)
 	while (s[cpt1])
 	{
 		cpt1 = cpt2;
-		while (s[cpt1] && cpt2 < (size_t)ft_strlen(s) && s[cpt2] != c)
+		while (s[cpt1] && cpt2 < ft_strlen(s) && s[cpt2] != c)
 			cpt2++;
 		tmp = ft_strsub(s, cpt1, cpt2 - cpt1);
 		if (tmp != NULL && (cpt2 - cpt1) > 0)
